# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 명장쇼핑몰 API
  description: 클라이언트 개발용 REST API 명세
  version: 1.0.0
servers:
  - url: http://localhost:3300
    description: 로컬 개발
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { type: string }
        phone: { type: string }
        role: { type: string, enum: [USER, ADMIN] }
        fishingPoints: { type: integer }
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: 회원가입
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 6 }
                name: { type: string }
                phone: { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
                  accessToken: { type: string }
                  refreshToken: { type: string }
                  expiresIn: { type: integer }
  /auth/login:
    post:
      tags: [Auth]
      summary: 로그인
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
                  accessToken: { type: string }
                  refreshToken: { type: string }
                  expiresIn: { type: integer }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: 토큰 갱신
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
                  accessToken: { type: string }
                  refreshToken: { type: string }
                  expiresIn: { type: integer }
  /auth/me:
    get:
      tags: [Auth]
      summary: 내 정보 조회
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
    patch:
      tags: [Auth]
      summary: 내 정보 수정
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 100 }
                phone: { type: string, maxLength: 20 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
  /auth/logout:
    post:
      tags: [Auth]
      summary: 로그아웃
      security: [{ bearerAuth: [] }]
      responses:
        '204':
          description: No Content
  /categories:
    get:
      tags: [Categories]
      summary: 카테고리 목록
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: isActive
          in: query
          schema: { type: boolean }
        - name: page
          in: query
          schema: { type: integer }
        - name: take
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: OK
    post:
      tags: [Categories]
      summary: 카테고리 생성
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 100 }
                description: { type: string }
                isActive: { type: boolean }
      responses:
        '201':
          description: Created
  /categories/{id}:
    get:
      tags: [Categories]
      summary: 카테고리 단건 조회
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
    patch:
      tags: [Categories]
      summary: 카테고리 수정
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                isActive: { type: boolean }
      responses:
        '200':
          description: OK
    delete:
      tags: [Categories]
      summary: 카테고리 삭제
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: No Content
...
  /cart:
    get:
      tags: [Cart]
      summary: 장바구니 조회
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
    post:
      tags: [Cart]
      summary: 장바구니 추가
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [productId, quantity]
              properties:
                productId: { type: string, format: uuid }
                optionId: { type: string, format: uuid }
                quantity: { type: integer, minimum: 1 }
      responses:
        '201':
          description: Created
    delete:
      tags: [Cart]
      summary: 장바구니 비우기
      security: [{ bearerAuth: [] }]
      responses:
        '204':
          description: No Content
  /cart/{itemId}:
    patch:
      tags: [Cart]
      summary: 장바구니 수량 수정
      security: [{ bearerAuth: [] }]
      parameters:
        - name: itemId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [quantity]
              properties:
                quantity: { type: integer, minimum: 1 }
      responses:
        '200':
          description: OK
    delete:
      tags: [Cart]
      summary: 장바구니 항목 삭제
      security: [{ bearerAuth: [] }]
      parameters:
        - name: itemId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: No Content
...
  /consulting:
    post:
      tags: [Consulting]
      summary: 컨설팅 신청
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [title, content]
              properties:
                title: { type: string, maxLength: 150 }
                content: { type: string }
      responses:
        '201':
          description: Created
  /consulting/me:
    get:
      tags: [Consulting]
      summary: 내 컨설팅 목록
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
...
  /orders:
    post:
      tags: [Orders]
      summary: 회원 주문 생성
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [items, paymentMethod]
              properties:
                items:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [productId, quantity]
                    properties:
                      productId: { type: string }
                      optionId: { type: string }
                      quantity: { type: integer, minimum: 1 }
                paymentMethod: { type: string }
                note: { type: string, maxLength: 255 }
      responses:
        '201':
          description: Created
  /orders/guest:
    post:
      tags: [Orders]
      summary: 비회원 주문 생성
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [guestName, guestEmail, items, paymentMethod]
              properties:
                guestName: { type: string, maxLength: 100 }
                guestEmail: { type: string, maxLength: 255 }
                guestPhone: { type: string, maxLength: 20 }
                items:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [productId, quantity]
                    properties:
                      productId: { type: string }
                      optionId: { type: string }
                      quantity: { type: integer, minimum: 1 }
                paymentMethod: { type: string }
                note: { type: string, maxLength: 255 }
      responses:
        '201':
          description: Created
  /orders/guest/{orderId}:
    get:
      tags: [Orders]
      summary: 비회원 주문 조회
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string }
        - name: email
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
  /orders/me:
    get:
      tags: [Orders]
      summary: 내 주문 목록
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
...